Refactor out BitnodesApi class to make it testable.
